+++ src/main/java/me/mrCookieSlime/Slimefun/api/energy/ChargableBlock.java
@@ -7,11 +7,12 @@ import org.bukkit.block.Block;
 import io.github.thebusybiscuit.cscorelib2.skull.SkullBlock;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.utils.HeadTexture;
-import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.api.Slimefun;
 
 public final class ChargableBlock {
+
+    private static final String KEY = "energy-charge";
 
     private ChargableBlock() {}
 
@@ -20,11 +21,12 @@ public final class ChargableBlock {
     }
 
     public static boolean isChargable(Location l) {
-        if (!BlockStorage.hasBlockInfo(l)) {
+        String id = BlockStorage.checkID(l);
+
+        if (id == null) {
             return false;
         }
 
-        String id = BlockStorage.checkID(l);
         return SlimefunPlugin.getRegistry().getEnergyCapacities().containsKey(id);
     }
 
@@ -33,13 +35,13 @@ public final class ChargableBlock {
     }
 
     public static int getCharge(Location l) {
-        String charge = BlockStorage.getLocationInfo(l, "energy-charge");
+        String charge = BlockStorage.getLocationInfo(l, KEY);
 
         if (charge != null) {
             return Integer.parseInt(charge);
         }
         else {
-            BlockStorage.addBlockInfo(l, "energy-charge", "0", false);
+            BlockStorage.addBlockInfo(l, KEY, "0", false);
             return 0;
         }
     }
@@ -60,17 +62,15 @@ public final class ChargableBlock {
             }
         }
 
-        if (charge != getCharge(l)) {
-            BlockStorage.addBlockInfo(l, "energy-charge", String.valueOf(charge), false);
-        }
+        BlockStorage.addBlockInfo(l, KEY, String.valueOf(charge), false);
     }
 
     public static void setUnsafeCharge(Location l, int charge, boolean updateTexture) {
         if (charge != getCharge(l)) {
-            BlockStorage.addBlockInfo(l, "energy-charge", String.valueOf(charge), false);
+            BlockStorage.addBlockInfo(l, KEY, String.valueOf(charge), false);
 
             if (updateTexture) {
-                updateCapacitor(l);
+                updateCapacitor(l, charge, getMaxCharge(l));
             }
         }
     }
@@ -79,42 +79,51 @@ public final class ChargableBlock {
         return addCharge(b.getLocation(), charge);
     }
 
-    public static int addCharge(Location l, int charge) {
-        int capacity = getMaxCharge(l);
-        int energy = getCharge(l);
-        int space = capacity - energy;
-        int rest = charge;
+    public static int addCharge(Location l, int addedCharge) {
+        String id = BlockStorage.checkID(l);
 
-        if (space > 0 && charge > 0) {
-            if (space > charge) {
-                setCharge(l, energy + charge);
+        if (id == null) {
+            BlockStorage.clearBlockInfo(l);
+            return 0;
+        }
+
+        int capacity = SlimefunPlugin.getRegistry().getEnergyCapacities().getOrDefault(id, 0);
+
+        int charge = getCharge(l);
+        int availableSpace = capacity - charge;
+        int rest = addedCharge;
+
+        if (availableSpace > 0 && addedCharge > 0) {
+            if (availableSpace > addedCharge) {
+                charge += addedCharge;
+                setCharge(l, charge);
                 rest = 0;
             }
             else {
-                rest = charge - space;
-                setCharge(l, capacity);
+                rest = addedCharge - availableSpace;
+                charge = capacity;
+                setCharge(l, charge);
             }
 
-            if (SlimefunPlugin.getRegistry().getEnergyCapacitors().contains(BlockStorage.checkID(l))) {
-                updateCapacitor(l);
+            if (SlimefunPlugin.getRegistry().getEnergyCapacitors().contains(id)) {
+                updateCapacitor(l, charge, capacity);
             }
         }
-        else if (charge < 0 && energy >= -charge) {
-            setCharge(l, energy + charge);
+        else if (addedCharge < 0 && charge >= -addedCharge) {
+            charge += addedCharge;
+            setCharge(l, charge);
 
-            if (SlimefunPlugin.getRegistry().getEnergyCapacitors().contains(BlockStorage.checkID(l))) {
-                updateCapacitor(l);
+            if (SlimefunPlugin.getRegistry().getEnergyCapacitors().contains(id)) {
+                updateCapacitor(l, charge, capacity);
             }
         }
 
         return rest;
     }
 
-    private static void updateCapacitor(Location l) {
+    private static void updateCapacitor(Location l, int charge, int capacity) {
         Slimefun.runSync(() -> {
             Block b = l.getBlock();
-            int charge = getCharge(b);
-            int capacity = getMaxCharge(b);
 
             if (b.getType() == Material.PLAYER_HEAD || b.getType() == Material.PLAYER_WALL_HEAD) {
                 if (charge < (int) (capacity * 0.25)) {
@@ -138,8 +147,7 @@ public final class ChargableBlock {
     }
 
     public static int getMaxCharge(Location l) {
-        Config cfg = BlockStorage.getLocationInfo(l);
-        String id = cfg.getString("id");
+        String id = BlockStorage.checkID(l);
 
         if (id == null) {
             BlockStorage.clearBlockInfo(l);